backend/requirements.txt
fastapi
uvicorn
gunicorn
openai
pinecone
sqlalchemy
asyncpg
python-dotenv
pyngrok
